requirements.txt
torch==1.7.1
torchvision==0.8.2
numpy==1.19.0
pytest==6.2.0
flake8==3.8.4
hydra-core==1.0.4
mypy==0.790
isort==5.6.4
